feat: Add GRPO and Dr.GRPO algorithms with CLI entry points and configurations.
